# specs/api.yaml
openapi: 3.1.0
info:
  title: PsiZero Resonance Platform API
  version: 1.0.0
  description: >
    Integrated APIs for symbolic resonance technologies:
    - SRS (Symbolic Resonance Solver) for NP-complete problems
    - HQE (Holographic Quantum Encoder) simulator
    - QSEM (Quantum Semantics) for concept resonance
    - NLC (Non-Local Communication) prime-eigenstate channels
    - QCR (Quantum Consciousness Resonator) sessions
    - I-CHING oracle (hexagram evolution)
    - Unified resonance physics (observer-capacity/gravity model)
servers:
  - url: https://api.psizero.dev
    description: Production
  - url: https://sandbox.psizero.dev
    description: Sandbox
...
security:
  - ApiKeyAuth: []
  - OAuth2CC: [platform.read, platform.write]
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
    OAuth2CC:
      type: oauth2
      description: Client Credentials flow (for enterprises)
      flows:
        clientCredentials:
          tokenUrl: https://auth.psizero.dev/oauth/token
          scopes:
            platform.read: Read access
            platform.write: Write/access compute
...
  responses:
    ProblemJson:
      description: RFC 7807 Problem Details
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [type, title]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string, format: uri }
        code: { type: string }
        meta: { type: object, additionalProperties: true }
...
    GravityRequest:
      type: object
      properties:
        observerEntropyReductionRate: { type: number, description: bits/sec (internal) }
        regionEntropyGradient: { type: number, description: bits/m^3 }
        blackHoleProxyDensity: { type: number, description: informational density proxy }
      required: [observerEntropyReductionRate]
    GravityResponse:
      type: object
      properties:
        effectiveG: { type: number, description: emergent gravitational constant (model units) }
        fieldStrength: { type: number }
        notes: { type: string }
paths:
...
  /v1/unified/gravity/compute:
    post:
      tags: [Unified]
      summary: Compute emergent gravity from observer/entropy parameters
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GravityRequest' }
            examples:
              eg:
                value:
                  observerEntropyReductionRate: 12.4
                  regionEntropyGradient: 0.002
      responses:
        '200':
          description: Effective gravitational outputs
          content:
            application/json:
              schema: { $ref: '#/components/schemas/GravityResponse' }
        '400': { $ref: '#/components/responses/ProblemJson' }

  # ---------- Webhooks ----------
...